requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.23
albumentations==1.4.24
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
av==15.0.0
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cnocr==2.3.2.1
cnstd==1.2.6.1
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
datasets==2.14.4
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
# Editable install with no version control (detectron2==0.6)
-e /Users/ayushraj/Documents/Python/WorkableAI/detectron2
dill==0.3.7
distro==1.9.0
doclayout_yolo==0.0.4
easyocr==1.7.2
einops==0.8.1
entmax==1.3
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.9
ipykernel==6.30.0
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lark==1.2.2
layoutparser==0.3.4
lazy_loader==0.4
lightning-utilities==0.15.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
munch==4.0.0
mypy_extensions==1.1.0
narwhals==2.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.19.2
onnxruntime-silicon==1.16.3
openai==1.97.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optimum==1.26.1
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
pix2tex==0.1.4
pix2text==1.1.4
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.0
pyspellchecker==0.8.3
pytesseract==0.3.13
pytest==8.4.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
rapidocr==3.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.33.2
shapely==2.0.7
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
streamlit==1.47.1
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
thop==0.1.1.post2209072238
tifffile==2024.8.30
tiktoken==0.9.0
timm==0.5.4
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.170
ultralytics-thop==2.0.14
Unidecode==1.4.0
uri-template==1.3.0
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
x-transformers==0.15.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
